requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
httpx[socks]==0.27.0
python-telegram-bot==21.3
beautifulsoup4==4.12.3
lxml==5.2.2
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.5
fake-useragent==1.5.1
tenacity==8.3.0
loguru==0.7.2
playwright==1.44.0
apscheduler==3.10.4
pydantic==2.7.4
pydantic-settings==2.3.4
